fix some regression bugs

+++ server/src/main/java/org/apache/iotdb/db/qp/executor/PlanExecutor.java
@@ -317,14 +317,17 @@ public class PlanExecutor implements IPlanExecutor {
   }
 
   private void flushSpecifiedStorageGroups(FlushPlan plan) throws StorageGroupNotSetException {
-    Map<PartialPath, List<Pair<Long, Boolean>>> storageGroupMap = plan.getStorageGroupPartitionIds();
+    Map<PartialPath, List<Pair<Long, Boolean>>> storageGroupMap = plan
+        .getStorageGroupPartitionIds();
     for (Entry<PartialPath, List<Pair<Long, Boolean>>> entry : storageGroupMap.entrySet()) {
       PartialPath storageGroupName = entry.getKey();
       // normal flush
       if (entry.getValue() == null) {
         if (plan.isSeq() == null) {
-          StorageEngine.getInstance().closeStorageGroupProcessor(storageGroupName, true, plan.isSync());
-          StorageEngine.getInstance().closeStorageGroupProcessor(storageGroupName, false, plan.isSync());
+          StorageEngine.getInstance()
+              .closeStorageGroupProcessor(storageGroupName, true, plan.isSync());
+          StorageEngine.getInstance()
+              .closeStorageGroupProcessor(storageGroupName, false, plan.isSync());
         } else {
           StorageEngine.getInstance()
               .closeStorageGroupProcessor(storageGroupName, plan.isSeq(), plan.isSync());
@@ -1069,12 +1072,21 @@ public class PlanExecutor implements IPlanExecutor {
     List<PartialPath> deletePathList = deleteTimeSeriesPlan.getPaths();
     try {
       List<String> failedNames = new LinkedList<>();
+      List<String> nonExistPaths = new ArrayList<>();
       for (PartialPath path : deletePathList) {
+        String failedTimeseries = "";
         StorageEngine.getInstance().deleteTimeseries(path.getDevicePath(), path.getMeasurement());
-        String failedTimeseries = IoTDB.metaManager.deleteTimeseries(path);
+        try {
+          failedTimeseries = IoTDB.metaManager.deleteTimeseries(path);
+        } catch (PathNotExistException e) {
+          nonExistPaths.add(path.getFullPath());
+        }
         if (!failedTimeseries.isEmpty()) {
           failedNames.add(failedTimeseries);
         }
+      }
+      if (!nonExistPaths.isEmpty()) {
+        throw new PathNotExistException(nonExistPaths);
       }
       if (!failedNames.isEmpty()) {
         throw new DeleteFailedException(String.join(",", failedNames));
